requirements.txt
flask
experta
fuzzywuzzy
python-Levenshtein
tensorflow==2.13.0
numpy==1.23.5
pillow==10.0.1
uuid
